refactor(uniffi): extract fire-uniffi-search
Phase 3a of the uniffi multi-namespace split.

Rust:
- Add rust/crates/fire-uniffi-search with its own
  setup_scaffolding!("fire_uniffi_search"). Hosts FireSearchHandle and the
  search/tag/user-mention record types previously in state_search.rs. The
  three async methods (search, search_tags, search_users) move off
  FireAppCore verbatim.
- FireAppCore gains a search: Arc<FireSearchHandle> field + search()
  getter. state_search.rs is deleted.
- Promote RequiredTagGroupState into fire-uniffi-types because it is
  consumed by both the new search crate (TagSearchResultState) and the
  session-domain TopicCategoryState (not yet extracted). state_session.rs
  imports it from the types crate now.

iOS native:
- FireSessionStore.swift routes search/searchTags/searchUsers through
  core.search() to the new FireSearchHandle. No Android call sites exist
  for these methods.
- project.yml preBuildScript outputs gain fire_uniffi_search.swift.
  Regenerated Fire.xcodeproj/project.pbxproj.

cargo test --workspace passes.
