+++ native/ios-app/project.yml
@@ -109,6 +109,7 @@ targets:
         outputFiles:
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_diagnostics.swift
+          - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_search.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_types.swift
           - $(SRCROOT)/Generated/fire_uniffiFFI/fire_uniffiFFI.h
           - $(SRCROOT)/Generated/fire_uniffiFFI/module.modulemap
